Push down scaffold into Fragments and hoist up ScaffoldState into ViewModel
