+++ Jetchat/app/src/main/java/com/example/compose/jetchat/profile/ProfileFragment.kt
@@ -29,7 +29,9 @@ import androidx.compose.ui.platform.ComposeView
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
+import androidx.navigation.findNavController
 import com.example.compose.jetchat.MainViewModel
+import com.example.compose.jetchat.components.JetchatScaffold
 import com.example.compose.jetchat.theme.JetchatTheme
 import com.google.accompanist.insets.LocalWindowInsets
 import com.google.accompanist.insets.ViewWindowInsetObserver
@@ -62,23 +64,25 @@ class ProfileFragment : Fragment() {
         val windowInsets = ViewWindowInsetObserver(this).start()
 
         setContent {
-            SideEffect {
-                activityViewModel.complexTopBar = false
-            }
+            JetchatScaffold(
+                scaffoldState = activityViewModel.scaffoldState,
+                complexTopBar = false,
+                findNavController = ::findNavController
+            ) {
+                val userData by viewModel.userData.observeAsState()
 
-            val userData by viewModel.userData.observeAsState()
-
-            CompositionLocalProvider(LocalWindowInsets provides windowInsets) {
-                JetchatTheme {
-                    if (userData == null) {
-                        ProfileError()
-                    } else {
-                        ProfileScreen(
-                            userData = userData!!,
-                            onNavIconPressed = {
-                                activityViewModel.openDrawer()
-                            }
-                        )
+                CompositionLocalProvider(LocalWindowInsets provides windowInsets) {
+                    JetchatTheme {
+                        if (userData == null) {
+                            ProfileError()
+                        } else {
+                            ProfileScreen(
+                                userData = userData!!,
+                                onNavIconPressed = {
+                                    activityViewModel.openDrawer()
+                                }
+                            )
+                        }
                     }
                 }
             }
